Updates connections
