+++ server/config/connection.js
@@ -1,13 +1,5 @@
 const mongoose = require('mongoose');
 
-const connectDB = async () => {
-  try {
-    await mongoose.connect(process.env.MONGODB_URI || 'mongodb://127.0.0.1:27017/dealership');
-    console.log('MongoDB connected successfully');
-  } catch (error) {
-    console.error('MongoDB connection error:', error);
-    process.exit(1);
-  }
-};
+mongoose.connect(process.env.MONGODB_URI || 'mongodb://127.0.0.1:27017/dealership');
 
-module.exports = connectDB;
+module.exports = mongoose.connection;
