+++ server/server.js
@@ -2,7 +2,7 @@
 const express = require('express');
 const { ApolloServer } = require('apollo-server-express');
 const path = require('path');
-const connectDB = require('./config/connection');
+const connectDB = require('./config/db');
 const typeDefs = require('./schemas/typeDefs');
 const resolvers = require('./schemas/resolvers');
 const authMiddleware = require('./utils/auth');
